requirements.txt
google-cloud-pubsub
discord-webhook